Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

Yes, while in the "a" and "ab" versions of the manual only the id_aa64pfr0_el1.MPAM field is indicated to assess the presence of the extension, in the most recent one ("ba") the id_aa64pfr1_el1.MPAM_frac has been introduced. Hence, to be compliant with this last version, probably the function could also check for the MPAM=0, MPAM_frac=1 combination, as well as for the 1,0 and 1,1 ones.
Since the fieds were both res0 before their introduction, I suppose this doesn't mess up with the previous versions of the architecture.

Patch-set: 1
Reviewer: Gerrit User 1000380 <1000380@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>