Update patch set 3
Patch Set 3:

Is this patch going to be progressed?

I see https://lists.trustedfirmware.org/pipermail/tf-a/2020-April/000399.html was the last message in the mail thread. It seemed to delay the discussion further until v2.3 was released which it has now been.

We monitor for patches that have been inactive for over 21 days and send a gentle reminder. Eventually though we abandon patches that appear to have stalled.

Patch-set: 3
CC: Gerrit User 1000049 <1000049@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>