Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Email sent to Scott who has passed it on to others in the Broadcom team on the various strategies that could be used to bring back some capability in this area as part of the broader Broadcomm platform port upstreaming.
The Broadcom platform port has now been upstreamed and it does not address this request.

We monitor for patches that have been inactive for over 21 days. 
Eventually though we abandon patches that appear to have stalled which this one has. If this issues is desired to be revisited that can come back as new patches.

Patch-set: 1