FPGA Support: Add support for topology self-discovery
As secondary cores show up, they populate the topology
tree by increasing the count of CPUs for the cluster
where they belong as well as the thread count.
This has to be done after the primary CPU has
initialized the C runtime.

The primary core will be responsible for initializing
the count of clusters as well as power domains and
setting the boundaries for the maximum affinity levels
that an MPIDR can have on the system, avoiding this way
to calculate wrong CPUIDs.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I8fd64761a2296714ce0f37c46544f3e6f13b5f61