FPGA Support: Add support for topology self-discovery
As secondary cores show up, they populate the topology
tree by increasing the count of cores for the cluster
where they belong. This has to be done after the primary
core has initialized the C runtime.

The primary core will be responsible for initializing
the count of clusters as well as power domains.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I8fd64761a2296714ce0f37c46544f3e6f13b5f61