Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > Thanks for the responses. 
> > Maybe I didn't understand the implementation correctly. Can I ask what is the difference between powering down(TURN OFF) only a CPU using CPU_SUSPEND versus CPU_OFF APIs? I see t194 supports later but not former.
> 
> Thanks. C7 for CPU idle uses a wake mask along with a wake time. This feature is broken on the HW and cannot be used. For CPU OFF, we enter C7 state with no wake time or wake mask


Hi Varun, just a suggestion from Soby so we complete this thread:
"Can the PSCI_SUSPEND call silently downgrade to a WFI within the firmware without returning an error? This provides compatibility with linux running with an old DT."
Would you prefer merging the patch as is or consider above comment? Up to you.

Patch-set: 2
CC: Gerrit User 1000228 <1000228@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>