Update patch set 1
Patch Set 1:

Yes, it goes this way and, thus, returns an error on SYSTEM_OFF to the OS. That this is possible looks like an ATF bug of its own to me because PSCI does not foresee this function to ever return.

But this patch does not address that issue in general, it just resolves it for the target platform so that we get a simple message on the console that no power-off is implemented, rather than undefined behavior (some Linux versions can crash in all colors then).

Patch-set: 1
Reviewer: Gerrit User 1000306 <1000306@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>