Update patch set 2
Patch Set 2: Code-Review-1

> Patch Set 1:
> 
> > The porting guide mentions the mandatory functions to be implemented:
> > https://trustedfirmware-a.readthedocs.io/en/latest/getting_started/porting-guide.html#plat-psci-ops-system-off
> > 
> > At least if not implemented, the service returns in a controlled manner rather than deferencing an undefined pointer.
> > Thus it's more a miss in the platform port than a TF-A bug. Never mind.
> 
> I still consider it a native ATF bug to allow a platform to return from no-return PSCI calls. That would be avoidable without relying on convention.

If the platform does not support this or the function fails it should still return IMHO. That way the system can handle the issue appropriately. Locking up the CPU in ATF is not always the right thing when a platform does not support hardware shutdown from ATF, a fallback may be in use that can shutdown the system. For instance Linux's EFI support seems to handle this return more gracefully[0] where it seems the crash you are seeing if from ARM's PSCI driver here[1].

If we really think this call should never return then it should be handled generically for all platforms that don't support this, vs a stub like this in each platform port.

[0]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/firmware/efi/reboot.c#n54

[1]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/firmware/psci/psci.c#n285

Patch-set: 2
Reviewer: Gerrit User 1000106 <1000106@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>
Label: Code-Review=-1