Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > > Patch Set 1:
> > > 
> > > > The porting guide mentions the mandatory functions to be implemented:
> > > > https://trustedfirmware-a.readthedocs.io/en/latest/getting_started/porting-guide.html#plat-psci-ops-system-off
> > > > 
> > > > At least if not implemented, the service returns in a controlled manner rather than deferencing an undefined pointer.
> > > > Thus it's more a miss in the platform port than a TF-A bug. Never mind.
> > > 
> > > I still consider it a native ATF bug to allow a platform to return from no-return PSCI calls. That would be avoidable without relying on convention.
> > 
> > If the platform does not support this or the function fails it should still return IMHO. That way the system can handle the issue appropriately. Locking up the CPU in ATF is not always the right thing when a platform does not support hardware shutdown from ATF, a fallback may be in use that can shutdown the system. For instance Linux's EFI support seems to handle this return more gracefully[0] where it seems the crash you are seeing if from ARM's PSCI driver here[1].
> > 
> > If we really think this call should never return then it should be handled generically for all platforms that don't support this, vs a stub like this in each platform port.
> > 
> > [0]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/firmware/efi/reboot.c#n54
> > 
> > [1]https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/firmware/psci/psci.c#n285
> 
> Returning an error from SYSTEM_OFF (and possibly some more functions like SYSTEM_RESET) is a PSCI spec violation. How an OS handles that or not does not matter, it remains a bug o the implementation by ATF. Yes, it is the responsibility of the platform to avoid stressing the generic code path that does this, but it is still a bug in the firmware.

I interpret the statement "should not return" as meaning it will not asynchronously start the shutdown sequence but instead it will wait for the power cut or reset.

If that is not the case then the PSCI spec is overly optimistic here, it also states the function will shutdown the system too, but that isn't always going to happen either, and in the case it does not work, returning at least allows the system to try to recover in some way.

Unlike consumer devices, a lot of embedded/industrial/automotive systems we target simply do not allow powering down from software, a safety reset is the best you get, so having a mandatory PSCI_SYSTEM_OFF is a problem with the spec in itself IMHO..

Anyway, adding a fake shutdown stub to the K3 platform code means we will advertise that we *do* have a way to shutdown the system from ATF when the HLOS checks with PSCI_FEATURES. What should happen is Linux (or whatever HLOS) should check for this call first using PSCI_FEATURES then forbid powerdown sequence if it's not available. I can work that angle in the kernel if you would like. We can also fixup ATF such that if this is still called then here[0] we don't simply comment but instead panic and halt in ATF instead of returning.

[0]https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/tree/lib/psci/psci_system_off.c#n33

Patch-set: 2
Reviewer: Gerrit User 1000106 <1000106@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>