Update patch set 2
Patch Set 2:

> > Returning an error from SYSTEM_OFF (and possibly some more functions like SYSTEM_RESET) is a PSCI spec violation. How an OS handles that or not does not matter, it remains a bug o the implementation by ATF. Yes, it is the responsibility of the platform to avoid stressing the generic code path that does this, but it is still a bug in the firmware.
> 
> I interpret the statement "should not return" as meaning it will not asynchronously start the shutdown sequence but instead it will wait for the power cut or reset.
> 

The spec (1.1) is clear: "Does not return".

> If that is not the case then the PSCI spec is overly optimistic here, it also states the function will shutdown the system too, but that isn't always going to happen either, and in the case it does not work, returning at least allows the system to try to recover in some way.
> 
> Unlike consumer devices, a lot of embedded/industrial/automotive systems we target simply do not allow powering down from software, a safety reset is the best you get, so having a mandatory PSCI_SYSTEM_OFF is a problem with the spec in itself IMHO..
> 

If you have concerns about the spec, discuss it with the authorities managing it. But this has been this way for multiple releases, and ATF or some OS is likely not the place to change it willingly.

> Anyway, adding a fake shutdown stub to the K3 platform code means we will advertise that we *do* have a way to shutdown the system from ATF when the HLOS checks with PSCI_FEATURES. What should happen is Linux (or whatever HLOS) should check for this call first using PSCI_FEATURES then forbid powerdown sequence if it's not available. I can work that angle in the kernel if you would like. We can also fixup ATF such that if this is still called then here[0] we don't simply comment but instead panic and halt in ATF instead of returning.
> 

Fixing the kernel to handle this ATF bug more gracefully would likely be good as well. Maybe it's better in upstream already, could migrate yet because we are waiting for you to finish that work.

But nothing is changed: PSCI demands no return from a this function. And so a lot of platform implementations already have stubs, now just one more. Panicing is what some do, but that is not clean either because it floods the console with - in this context - useless information.

Patch-set: 2
Reviewer: Gerrit User 1000306 <1000306@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>