Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2:
> 
> > > Returning an error from SYSTEM_OFF (and possibly some more functions like SYSTEM_RESET) is a PSCI spec violation. How an OS handles that or not does not matter, it remains a bug o the implementation by ATF. Yes, it is the responsibility of the platform to avoid stressing the generic code path that does this, but it is still a bug in the firmware.
> > 
> > I interpret the statement "should not return" as meaning it will not asynchronously start the shutdown sequence but instead it will wait for the power cut or reset.
> > 
> 
> The spec (1.1) is clear: "Does not return".
> 
> > If that is not the case then the PSCI spec is overly optimistic here, it also states the function will shutdown the system too, but that isn't always going to happen either, and in the case it does not work, returning at least allows the system to try to recover in some way.
> > 
> > Unlike consumer devices, a lot of embedded/industrial/automotive systems we target simply do not allow powering down from software, a safety reset is the best you get, so having a mandatory PSCI_SYSTEM_OFF is a problem with the spec in itself IMHO..
> > 
> 
> If you have concerns about the spec, discuss it with the authorities managing it. But this has been this way for multiple releases, and ATF or some OS is likely not the place to change it willingly.

ATF is essentially the reference implementation, we should focus on doing the correct thing in code, and patch the spec if needed. Many of the managing authorities are here working on TF-A.

But I'm also not arguing changing the spec, simply pointing out a shortcoming I am seeing.

> 
> > Anyway, adding a fake shutdown stub to the K3 platform code means we will advertise that we *do* have a way to shutdown the system from ATF when the HLOS checks with PSCI_FEATURES. What should happen is Linux (or whatever HLOS) should check for this call first using PSCI_FEATURES then forbid powerdown sequence if it's not available. I can work that angle in the kernel if you would like. We can also fixup ATF such that if this is still called then here[0] we don't simply comment but instead panic and halt in ATF instead of returning.
> > 
> 
> Fixing the kernel to handle this ATF bug more gracefully would likely be good as well. Maybe it's better in upstream already, could migrate yet because we are waiting for you to finish that work.

Upstream is the same as 5.4 in this area. Plus AM65x is functional upstream (outside of DSP and Graphics, both we are working to remedy) if you would like to test this.

I also cant fix this in kernel if we advertise we have a PSCI_SYSTEM_OFF function even though we do not have a working one.

> 
> But nothing is changed: PSCI demands no return from a this function. And so a lot of platform implementations already have stubs, now just one more. Panicing is what some do, but that is not clean either because it floods the console with - in this context - useless information.

Game plan for K3 is to have this SYSTEM_OFF function call into our central system controller (DMSC) to actually provide the PMIC powerdown if allowed (we may have MCU side active and the main domain should not be able to powerdown the MCU side, that would be a safety isolation violation).

Panic() is probably the correct thing to do here for now, but I'm assuming the point of this patch is to get rid of all that console spam and errors we currently have when trying to powerdown, so this simple message is okay also.

Since we don't have a real powerdown just yet, I cant say this is any less correct than what we have now so I wont block it.

Patch-set: 2
Reviewer: Gerrit User 1000106 <1000106@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>
Label: Code-Review=+1