Update patch set 2
Patch Set 2:

Just my view on this,

SYSTEM_OFF is one of the mandatory functions of PSCI. The PSCI implementation (TF-A) allows platforms to choose to not implement one of the mandatory functions and will return SMC_UNK for a non implemented PSCI function. What this means is that such a platform is not fully PSCI spec compliant anymore since if it chooses to not implement a mandatory function.

This approach also allows platforms to bootstrap PSCI support without having to mandate full PSCI compliance from day 1. So as far as compliance to PSCI spec goes, TF-A allows the platforms to be non-compliant. This can be seen as a bug or flexibility depending on how you look at it.

I can see that the firmware could provide a default SYSTEM_OFF for PSCI, but as already mentioned - Locking up the CPU in TF-A is not always the right thing when a platform does not support hardware shutdown from TF-A, other fallback options might be available to the platform.

TF-A shouldn't force a default implementation for SYSTEM_OFF IMHO. The decision should be left to the platform.

Patch-set: 2
CC: Gerrit User 1000056 <1000056@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>