Tegra: memctrl_v2: fixup sequence to resize video memory
The previous sequence used by the driver to program the new memory aperture
settings and clear the non-overlapping memory was faulty. The sequence locked
the non-overlapping regions twice, leading to faults when trying to clear it.

This patch modifies the sequence to follow these steps:

* move the previous memory region to a new firewall register
* program the new memory aperture settings
* clean the non-overlapping memory

This patch also maps the non-overlapping memory as Device memory to follow
guidance from the arch. team.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I7cf6e05b2dd372103dc7229e37b1b3fc269a57ae