arm_fpga: Add support to populate the CPU nodes in the DTB
At the moment BL31 dynamically discovers the CPU topology of an FPGA
system at runtime, but does not export it to the non-secure world.
Any BL33 user would typically look at the devicetree to learn about
existing CPUs.

Add generic code to create and populate the /cpus node in a devicetree.
We use just the minimum required nodes and properties to satisfy the
binding, so we don't describe a cpumap or caches at the moment (this
gets hairy with the references, but can be added later).

An existing /cpus node in the DT will make the code bail out with a
message. So any DT that is supposed to be filled must lack this node.
Also this means that the topology can be forced in the DT if needed.

Also call this new function from the FPGA port.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I589a2b3412411a3660134bdcef3a65e8200e1d7e