drivers: allwinner: axp: Enable regulators more judiciously
Some regulators should not be enabled at boot, even if they have a
phandle, such as when they have timing or sequencing requirements.

For example, the Ethernet PHY on the Orange Pi 3 has two power supplies:
one from the PMIC and one from a gpio-controlled regulator. The
gpio-controlled regulator must be enabled for the PHY to initialize
properly. The code here breaks that.

Regulators should need to opt in to being enabled at boot.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Change-Id: Iffd4f06b5d4806d4636d73f960c129140ba34bcb