drivers: allwinner: axp: Do not force DRIVEVBUS on
It is not necessarily safe to enable DRIVEVBUS at boot. For example,
in a battery-powered device where VBUS is used for charging, enabling
DRIVEVBUS at boot causes a loop and prevents charging from a
dead-battery condition.

Furthermore, since bit 4 of register 0x8f is cleared by default and is
persisted across reboots, it is never useful to explicitly clear it.
For this reason, remove the whole code block.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Change-Id: I26d75b229f235e8c5630e75aadcfc02dc9ca24b2