Update patch set 1
Patch Set 1:

> Hi, have you seen https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/d93eb4467a0ca09d359f242b19f584a58e4e805b%5E%21/#F0 ?

No, I hadn't, but even that commit message mentions that this is the wrong thing to do for micro-B ports.

> So DRIVEVBUS was explicitly enabled to fix a problem. Can you elaborate on that loop you mentioned? And what device does this affect?

It causes a problem for any device where USB VBUS is the external power source. This is known to affect the PinePhone,

The loop happens when the battery is dead or removed. In that case, the device id drawing power from VBUS when it is turned on. Then ATF runs, sees x-powers,drive-vbus-en, and turns on DRIVEVBUS. This cuts off the VBUS power source path, and the device dies as soon as the capacitors drain. When the PMIC loses power, it stops blocking VBUS. This causes the PMIC to receive power again and turn the phone back on. ATF runs, turns on DRIVEVBUS, which turns off VBUS, etc., etc.

Patch-set: 1