Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Looks harmless enough, but is there any reason you need this?
> Do you have a system which actually honours the SPC (having the secure boot fuse burnt)?
> Wouldn't you need to switch the whole PRCM control in the SPC to secure as well?
> Also, where did you find those bits explained? I just copied the line from the BSP ATF, and the A80 manual (which has the PRCM registers description) is lacking info about this exact register.

The reason is to prevent someone from writing a Linux driver that touches the VDD_SYS gating bits.

The security switch bits are honored even without the fuse burnt and without the PRCM set to secure in the SPC. Without this change, "devmem2 0x01f01510 w 0xf" will hang the system. With this change, it does nothing.

The bits are documented in the comment here (same as all of the other ATF dumps) and verified by experiment:
https://github.com/Allwinner-Homlet/H6-BSP4.9-brandy/blob/master/arm-trusted-firmware-1.0/plat/sun50iw1p1/sunxi_security.c

Patch-set: 1