Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Owner-Review-1
> 
> I am sorry, but this breaks almost every board in U-Boot, mostly for Ethernet and HDMI, but also USB on many boards. U-Boot relies on TF-A to enable those regulators.

I am well aware of this. However, I need a place to point people affected by the current broken code, so they at least have a workaround patch.

> What is the particular problem with the OPi-3? Does the GPIO power need to be enabled *before* ALDO2? Also I don't see Ethernet at all in Linux' mainline DTB, and no OPi-3 support in U-Boot at all?
> U-Boot can enable GPIO controlled regulators, just add CONFIG_MACPWR="PD6" in the defconfig.

The two PHY power rails need to be powered up at the same time. See [1] and the rest of the series. Ondrej said that even a 50ms time difference prevented the PHY from working.

[1]: https://patchwork.kernel.org/patch/11103967/

The root problem here is that it is impossible to exclude a regulator from being enabled. Either an opt-in or an opt-out device tree property would work, and I want to have a discussion about what that would look like. Until there is a way to control this code from the device tree, this problem will keep coming up.

Patch-set: 1