Update patch set 4
Patch Set 4:

Hi Madhu,

> I noticed that the libfdt source files now have new APIs which are not used by TF-A leading to increase in binary size. Ex: fdt_appendprop_addrrange()

If these new APIs are not used by TF-A, I would have expected the linker to prune them... Do you know why this is not the case?

> I checked the increase in text segment in BL31 and found that it significantly varies based on build configuration. One build config had approx 0x2f0 increase compared to 0x7b0 for the above fvp-tspd-tbb-mbedtls-pauth. However, this increase in text segment size together with the alignment restrictions of exception vectors leads to an additional 4K increase in size of BL31 image.

Ah I see, we've crossed a page boundary... Too bad.

> I have leveraged USE_ROMLIB feature to reduce the size of various segments in BL31 and successfully build TF-A images.

That's good news! I am not seeing this in this patch, will that come in version 5 of the patch or will it be part of a separate patch?

Patch-set: 4
Reviewer: Gerrit User 1000043 <1000043@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>