Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Owner-Review-1
> > > 
> > > I am sorry, but this breaks almost every board in U-Boot, mostly for Ethernet and HDMI, but also USB on many boards. U-Boot relies on TF-A to enable those regulators.
> > 
> > I am well aware of this. However, I need a place to point people affected by the current broken code, so they at least have a workaround patch.
> 
> Please mark it as an RFC patch or hack, then. We can't just have patches lying around that break everything.
> And while in an ideal world TF-A indeed wouldn't care about those regulators, this is the way we do it on this platform.
> If now this easy approach doesn't work anymore, we first need to fix this, for instance by adding RSB and AXP support in U-Boot. Which is frankly speaking quite some work, and it's a bummer that a single board requires this (but see below).
> 
> > > What is the particular problem with the OPi-3? Does the GPIO power need to be enabled *before* ALDO2? Also I don't see Ethernet at all in Linux' mainline DTB, and no OPi-3 support in U-Boot at all?
> > > U-Boot can enable GPIO controlled regulators, just add CONFIG_MACPWR="PD6" in the defconfig.
> > 
> > The two PHY power rails need to be powered up at the same time. See [1] and the rest of the series. Ondrej said that even a 50ms time difference prevented the PHY from working.
> 
> But can't Linux turn off the AXP rail before initiating the sequence? I guess it would need to do it anyway, when U-Boot gains support, and in general to make sure we start from a clean sheet.
> 
> > 
> > [1]: https://patchwork.kernel.org/patch/11103967/
> > 
> > The root problem here is that it is impossible to exclude a regulator from being enabled. Either an opt-in or an opt-out device tree property would work, and I want to have a discussion about what that would look like. Until there is a way to control this code from the device tree, this problem will keep coming up.
> 
> I understand, but I don't think we will get a kind of "exclude" property, as this is basically what not having "regulator-always-on;" means. If it's really only needed for one (or two) boards, we might investigate alternative solutions:
> - We detect in TF-A that there is a "sequence requirement" for a particular AXP rail, then refrain from activating it.
> - We actually support a sequenced regulator, and activate the GPIO along with it. We have GPIO support already for the H5 power down.
> - We hack an exception into TF-A, to skip this regulator on an OrangePi-3, based on the DT root model name.
> 
> Shall we discuss this more openly, on some mailing list? Is Ondrej working on it right now? I see that the series is from August last year.

It's not just Orange Pi 3, it's also on Oragnge Pi One Plus (which has the same design around the phy).

The kernel regulator consumer code can not disable the regulator that was not enabled by a consumer yet, due to reference counting. It can't even use regulator_force_disable, because that is also gated by reference count (+ there's no reverse operation, since it should only be used in emergency to shutdown the regulator) not being 0 + there's no reverse operation to re-enable the force disabled regulator. Also on sunxi, dwmac sunxi bridge only implements one regulator, and all the boards that need two regulators for the phy workaround this by using invalid HW description where one regulator is described as being an input supply for the other one, which makes it even more complicated, because you only get access to one of the regulators from the regulator consumer code.

In short Linux should ideally get a cleanly powered up PHY, to avoid all these issues.

Another thing is that on SoC reset, the gpio is cleared, but the AXP regulator is kept powered on.
This can be fixed in the ehthernet driver, by disabling PHY power first. Otherwise ethernet will not work even in u-boot, because u-boot/ATF doesn't power cycle both regulators either.

It's a real mess.

The best solution it seems is for ATF/u-boot to cleanup the PHY power state and not pass it in broken state further down to Linux in any case. U-boot can't assume PHY is powered up correctly anyway, due to the SoC reset clearing up gpio state anyway.

Patch-set: 1
CC: Gerrit User 1000282 <1000282@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>