Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> Hi, have you seen https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/d93eb4467a0ca09d359f242b19f584a58e4e805b%5E%21/#F0 ?
> So DRIVEVBUS was explicitly enabled to fix a problem. Can you elaborate on that loop you mentioned? And what device does this affect?

It's not problem making DRIVEVBUS (N_VBUSEN pin) an output. It's problem when we also start driving it to a high state without any further consideration for the particular board design, and in particular for the state of the ID pin, just based on x-powers,drive-vbus-en. Some boards don't even have an ID pin, and use USB-C controller to determine power role, and whether VBUS should be enabled or not.

It's not hard to imagine a board where this will enable the VBUS output to the USB port that will also clash with the charger that may be connected there. For example on TBS A711 tablet (doesn't use ATF though, just using it as an example where this would cause potentially backfeeding power to whatever's connected to the device's USB port), which is based on some reference Allwinner design, this is all that's needed to power up the VBUS. Some other devices, like Pinephone need to enable the 5V regulator separately first, for VBUS to have an actual power, but it's not something that can be assumed.

Patch-set: 1
CC: Gerrit User 1000282 <1000282@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>