Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > (2 comments)
> > > 
> > > Hi, have you seen https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/d93eb4467a0ca09d359f242b19f584a58e4e805b%5E%21/#F0 ?
> > > So DRIVEVBUS was explicitly enabled to fix a problem. Can you elaborate on that loop you mentioned? And what device does this affect?
> > 
> > It's not problem making DRIVEVBUS (N_VBUSEN pin) an output. It's problem when we also start driving it to a high state without any further consideration for the particular board design, and in particular for the state of the ID pin, just based on x-powers,drive-vbus-en. Some boards don't even have an ID pin, and use USB-C controller to determine power role, and whether VBUS should be enabled or not.

Yes, I understand that it's not just about the pinmux. I made a small patch to properly separate the two, but this doesn't matter, as we would always switch both together.
So as the commit message referenced above says, I am well aware of this not being the ideal solution, but so far it was working quite well. So apparently now we need a new solution, but one that doesn't break existing boards. And preferably one that it is as easy as possible.

> > It's not hard to imagine a board where this will enable the VBUS output to the USB port that will also clash with the charger that may be connected there. For example on TBS A711 tablet (doesn't use ATF though, just using it as an example where this would cause potentially backfeeding power to whatever's connected to the device's USB port), which is based on some reference Allwinner design, this is all that's needed to power up the VBUS. Some other devices, like Pinephone need to enable the 5V regulator separately first, for VBUS to have an actual power, but it's not something that can be assumed.
> 
> What I'm trying to say is that current behavior doesn't makes sense and is not safe for poretable devices. The default should be to boot with VBUS disabled and only enable it when the power role for the USB port is clearly determined, either by u-boot or by a Linux driver.

Sure, but U-Boot cannot enable AXP rails over RSB, because it has no driver for it. The simple solution so far was just to enable all regulators that we *might* need, in TF-A. I see that this approach lacks control, but we can't just disable this code without finding an alternative solution.

> And even in u-boot the users most likely want the device to act as a peripheral, not as USB host.

That's debatable, for booting distributions (installers) USB keyboard, USB network adapters and USB flash drives are far more useful.

So that being said: I will try to check all supported boards in this manner, maybe we can indeed turn this code off by default, but have a build time option to enable this for the odd boards that need it (Remix Mini). Even though I would very much keep the single TF-A build for all boards.

Patch-set: 1
Reviewer: Gerrit User 1000118 <1000118@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>